feat: start command integration
