+++ src/config.ts
@@ -1,2 +1,2 @@
-export const TELEGRAM_BOT_TOKEN = process.env.TELEGRAM_BOT_TOKEN as string;
+export const ARSENAL_GAMES_TELEGRAM_BOT_TOKEN = process.env.ARSENAL_GAMES_TELEGRAM_BOT_TOKEN as string;
 export const LOG_LEVEL = 'debug';
