+++ src/telegram.ts
@@ -1,24 +1,25 @@
 import TelegramBot from 'node-telegram-bot-api';
-import { TELEGRAM_BOT_TOKEN } from './config';
+import { ARSENAL_GAMES_TELEGRAM_BOT_TOKEN } from './config';
 import { initLog } from './logger';
+import { scrapeArsenalFixtures } from './scrape';
 
 const logger = initLog('telegram');
 
-const bot = new TelegramBot(TELEGRAM_BOT_TOKEN, { polling: true });
+const bot = new TelegramBot(ARSENAL_GAMES_TELEGRAM_BOT_TOKEN, { polling: true });
 
-let chatId = 361786193;
-
-export function sendRichText(richText: string) {
+export function sendRichText(richText: string, chatId) {
   if (chatId) {
     bot.sendMessage(chatId, richText, { parse_mode: 'Markdown' });
+    logger.info(`[${chatId}] Message sent`);
   } else {
     logger.error('No chatId!');
   }
 }
 
 bot.onText(/\/start/, (msg) => {
-  chatId = msg.chat.id;
-  logger.info(chatId);
+  const chatId = msg.chat.id;
+  logger.info(`[${chatId}] Got start command!`);
+  scrapeArsenalFixtures(chatId);
 });
 
 bot.on('text', (msg) => {
